+++ MineCraft/Source/Util/Random.hpp
@@ -26,7 +26,7 @@ private:
 template<typename REngine = std::mt19937>
 class Random {
 public:
-    Random(int n = std::time(nullptr)) {
+    Random(int n = (int) std::time(nullptr)) {
         m_randomEngine.seed(n);
         for (int i = 0; i < 5; i++) {
             intInRange(i, i * 5);
